Run parsing on experimental

+++ pipeline/pak-tools/src/pak/structs/UScript/UScriptStrutTypes/FFrameNumber.ts
@@ -1,9 +1,9 @@
-import {Float} from "../../../primitive/decimals";
 import {Reader} from "../../../../readers/Reader";
+import {Int32} from "../../../primitive/integers";
 
 
 export async function FFrameNumber(reader: Reader) {
   return {
-    Value: await reader.read(Float),
+    Value: await reader.read(Int32),
   };
 }
